+++ server/services/patreon.service.js
@@ -15,6 +15,7 @@ async function grabLatestPatreonData() {
   if (latestPatreonData.status === 200) {
     const getJSON = await latestPatreonData.json();
     patreonData = getJSON.patreons.applicationPledges;
+
     logger.debug("Successfully grabbed remote patreon data!");
   } else {
     logger.error("Falling back to local patreon data...");